interface/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-debug-toolbar==3.7.0
django-ninja==0.21.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
executing==0.8.1
Faker==12.0.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
mixer==7.2.2
orjson==3.8.7
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.5
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.1
shortuuid==1.0.9
simple-print==1.6.14
six==1.16.0
sqlparse==0.4.3
termcolor==1.1.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.13
whitenoise==6.2.0
